refactor: test 에 있는 클래스 선언들을 main 으로 이동시켰다.
