+++ src/main/kotlin/resume/Person.kt
@@ -1,0 +1,6 @@
+package resume
+
+class Person(
+    val name: String,
+    val company: String?,
+)
