+++ src/test/kotlin/resume/ResumeTest.kt
@@ -41,25 +41,3 @@ class ResumeTest {
         person.company shouldBe "DOGE"
     }
 }
-
-fun introduce(init: PersonBuilder.() -> Unit): Person = PersonBuilder().apply(init).build()
-
-class Person(
-    val name: String,
-    val company: String?,
-)
-
-class PersonBuilder {
-    lateinit var name: String
-    var company: String? = null
-
-    fun name(value: String) {
-        name = value
-    }
-
-    fun company(value: String) {
-        company = value
-    }
-
-    fun build() = Person(name, company)
-}
